requirements.txt
requests==2.26.0
click==8.0.1
pyyaml==5.4.1